Move app.js to root
This is such an important init part of the app, it should be in root
